Update implementation of `since` in API to Unix timestamps
The `since` implementation in the API was changed to use Unix timestamps across the request-response lifecycle, instead of date strings/objects.

Earlier, frontend was sending the unix timestamp that concurred with the time instance the request was completed processing at the frontend, including update of the nodes & edges in the engine store.
This results in the backend not returning the correct nodes & edges, missing the ones created/updated between the time the request was sent and the time it was processed at the frontend.

Ideally, the request should send the maximum timestamp of the nodes/edges present in the current state of the store to ensure correct nodes & edges are returned.

===

While making this change, another issue was encountered while trying to send the Unix timestamp of `node.written_at` derived from the date string stored in the frontend, which misses the sub-millisecond value due to the lack of it's support in the `Date` object of all browsers.

The below debug message demonstrates an example of this issue:

```sh
Node written at: 2025-04-05 13:17:58.964821 UTC, Received since value: 2025-04-05 13:17:58.964 UTC
```

In this case, the `since` request was always returning that one last node, triggering a store update & a UI re-render for it each time inherently, even though the node was already present in the store.

===

The change in this commit achieves the following:
- It's the first step towards a uniform timestamp format across the request-response lifecycle.
- Enables frontend to store the `written_at` timestamps for all nodes/edges in the store directly received from the backend, eliminating the need to convert them to unix timestamps while forming the `since` parameter.
- Reduces total number of conversions needed when processing requests. The `since` parameter is now directly compared with the millisecond unix timestamp of `node.written_at` & `edge.written_at`, instead of converting it to a date object first.